+++ terraform/aws/dev/installers/installer.sh
@@ -62,6 +62,9 @@ SATISFACTORY_EXPERIMENTAL_FS_UUID="0fab4de1-05ee-4399-928f-ea4a4d1fd690"
 # Application install, update, configuration, and execution
 # -----
 
+echo "INFO: Create a tmp working dir to use as needed"
+mkdir -p "${INSTALLER_DIR_PATH}" || exit 1
+
 # -----
 # Factorio (~/factorio)
 # -----
@@ -84,10 +87,14 @@ then
 
     if [[ ! -f /home/ubuntu/factorio/bin/x64/factorio ]]
     then
-        echo "INFO: Installing Factorio dedicated server"
-        cd /usr/local/lanordie/gameserver/
-        tar -xJf factorio_headless_x64_1.1.74.tar.xz -C /home/ubuntu
-        cd /home/ubuntu/factorio/bin/x64 || exit 1
+        echo "INFO: Install Factorio archive"
+        cd "${INSTALLER_DIR_PATH}" || exit 1
+        mkdir -p /home/ubuntu/factorio || exit 1
+
+        aws s3 cp "s3://${BUCKET_ID}-${DEPLOYMENT_ID}/factorio_headless_x64_1.1.74.tar.xz" .
+        tar -xf factorio_headless_x64_1.1.74.tar.xz -C /home/ubuntu/pa_titans/stable
+
+        cd "${INSTALLER_DIR_PATH}"/
         ./factorio --create /home/ubuntu/factorio/saves/lanordie.zip
         ls -la /home/ubuntu/factorio/saves
     fi
@@ -132,72 +139,89 @@ fi
 # KSP (~/ksp)
 # -----
 
-# if [[ $ENABLE_KSP == true ]]
-# then
-#     echo "INFO: Mount Kerbal Space Program EBS volume"
-#     mkdir -p /home/ubuntu/ksp || true
+if [[ $ENABLE_KSP == true ]]
+then
+    echo "INFO: Mount Kerbal Space Program EBS volume"
+    mkdir -p /home/ubuntu/ksp || true
 
-#     # check if the FS is already listed in /etc/fstab
-#     # shellcheck disable=SC2143
-#     if [[ ! $(grep -rnw '/etc/fstab' -e "$KSP_FS_UUID") ]]
-#     then
-#         echo "UUID=$KSP_FS_UUID /home/ubuntu/ksp ext4 defaults,errors=remount-ro 0 1" >> /etc/fstab
-#         mount -a
-#     fi
+    # check if the FS is already listed in /etc/fstab
+    # shellcheck disable=SC2143
+    if [[ ! $(grep -rnw '/etc/fstab' -e "$KSP_FS_UUID") ]]
+    then
+        echo "UUID=$KSP_FS_UUID /home/ubuntu/ksp ext4 defaults,errors=remount-ro 0 1" >> /etc/fstab
+        mount -a
+    fi
 
-#     echo "INFO: /etc/fstab contents"
-#     cat /etc/fstab
+    echo "INFO: /etc/fstab contents"
+    cat /etc/fstab
 
-#     if [[ ! -f /home/ubuntu/factorio/bin/x64/factorio ]]
-#     then
-#         echo "INFO: Install KSP base"
-#         unzip -q ksp-linux-1.12.5.zip
-#         cp -rf KSP_linux/. /home/ubuntu/ksp/
-#         rm -rf KSP_linux
-#         echo "INFO: Install KSP expansions"
-#         unzip -q KSP-Breaking_Ground_Expansion-en-us-lin-1.7.1.zip -d /home/ubuntu/ksp
-#         unzip -q KSP-Making_History_Expansion-en-us-lin-1.12.1.zip -d /home/ubuntu/ksp
-#         cd /home/ubuntu/ksp
-#         ./dlc-bge-en-us.sh
-#         ./dlc-mhe-en-us.sh
+    if [[ ! -f /home/ubuntu/factorio/bin/x64/factorio ]]
+    then
+        echo "INFO: Install KSP base"
+        cd "${INSTALLER_DIR_PATH}" || exit 1
+        mkdir -p "/home/ubuntu/ksp/" || exit 1
 
-#         echo "INFO: Installing and update KSP DMPServer dedicated server and DMPUpdater"
-#     fi
+        echo "INFO: Copy KSP base archive"
+        aws s3 cp "s3://${BUCKET_ID}-${DEPLOYMENT_ID}/ksp-linux-1.12.5.zip" .
+        unzip -oqq ksp-linux-1.12.5.zip /home/ubuntu/ksp/
 
-#     echo "INFO: Resetting Kerbal Space Program dir ownership"
-#     chown ubuntu:ubuntu -R /home/ubuntu/ksp
+        echo "INFO: Install KSP expansions"
+        aws s3 cp "s3://${BUCKET_ID}-${DEPLOYMENT_ID}/KSP-Breaking_Ground_Expansion-en-us-lin-1.7.1.zip" .
+        unzip -oqq KSP-Breaking_Ground_Expansion-en-us-lin-1.7.1.zip -d /home/ubuntu/ksp
 
-#     if [[ ! -f "/etc/systemd/system/ksp.service" ]]
-#     then
+        aws s3 cp "s3://${BUCKET_ID}-${DEPLOYMENT_ID}/KSP-Making_History_Expansion-en-us-lin-1.12.1.zip" .
+        unzip -oqq KSP-Making_History_Expansion-en-us-lin-1.12.1.zip -d /home/ubuntu/ksp
 
-#         echo "INFO: Create system service for KSP"
-#         echo -e "\
-#         [Unit]
-#             After=syslog.target network.target nss-lookup.target network-online.target
-#             Description=KSP dedicated server
-#             Wants=network-online.target
+        cd /home/ubuntu/ksp
+        chmod +x ./dlc-bge-en-us.sh
+        ./dlc-bge-en-us.sh
+        chmod +x ./dlc-mhe-en-us.sh
+        ./dlc-mhe-en-us.sh
 
-#         [Service]
-#             ExecStart=mono /usr/share/ksp/DMPServer.exe
-#             Group=ubuntu
-#             KillSignal=SIGINT
-#             Restart=on-failure
-#             StandardOutput=journal
-#             User=ubuntu
-#             WorkingDirectory=/usr/share/ksp
+        echo "INFO: Installing and update KSP DMPServer dedicated server and DMPUpdater"
+        aws s3 cp "s3://${BUCKET_ID}-${DEPLOYMENT_ID}/DMPServer.zip" .
+        unzip -oqq DMPServer.zip -d /home/ubuntu/ksp
 
-#         [Install]
-#             WantedBy=multi-user.target" | tee "/etc/systemd/system/ksp.service"
+        cp -rf DMPModpackUpdater.exe /home/ubuntu/ksp
+        mono DMPModpackUpdater.exe
 
-#         sed -i 's/^ *//g' "/etc/systemd/system/ksp.service"
+        cd /home/ubuntu/ksp
+    fi
 
-#         systemctl enable ksp.service --now
-#         systemctl restart ksp.service
-#     fi
+    echo "INFO: Resetting Kerbal Space Program dir ownership"
+    chown ubuntu:ubuntu -R /home/ubuntu/ksp
 
-#     echo "INFO: Kerbal Space Program service status"
-#     systemctl status ksp.service
-# fi
+    if [[ ! -f "/etc/systemd/system/ksp.service" ]]
+    then
+
+        echo "INFO: Create system service for KSP"
+        echo -e "\
+        [Unit]
+            After=syslog.target network.target nss-lookup.target network-online.target
+            Description=KSP dedicated server
+            Wants=network-online.target
+
+        [Service]
+            ExecStart=mono /home/ubuntu/ksp/DMPServer/DMPServer.exe
+            Group=ubuntu
+            KillSignal=SIGINT
+            Restart=on-failure
+            StandardOutput=journal
+            User=ubuntu
+            WorkingDirectory=/home/ubuntu/ksp
+
+        [Install]
+            WantedBy=multi-user.target" | tee "/etc/systemd/system/ksp.service"
+
+        sed -i 's/^ *//g' "/etc/systemd/system/ksp.service"
+
+        systemctl enable ksp.service --now
+        systemctl restart ksp.service
+    fi
+
+    echo "INFO: Kerbal Space Program service status"
+    systemctl status ksp.service
+fi
 
 # -----
 # Planetary Annihilation : Titans (~/ps_titans)
@@ -224,7 +248,10 @@ then
     then
         # source https://store.planetaryannihilation.net/Download/Install?titleId=4&sessionTicket=7065387651658097800
         echo "INFO: Install Planetary Annihilation : Titans archive"
-        mkdir -p /home/ubuntu/pa_titans/stable
+        cd "${INSTALLER_DIR_PATH}" || exit 1
+        mkdir -p /home/ubuntu/pa_titans/stable || exit 1
+
+        aws s3 cp "s3://${BUCKET_ID}-${DEPLOYMENT_ID}/PA_Linux_116400.tar.bz2" .
         tar -xf PA_Linux_116400.tar.bz2 -C /home/ubuntu/pa_titans/stable
     fi
 
@@ -238,9 +265,9 @@ then
 
         # Key/Values are passed in via instance data.template_file.user_data invocation
         # shellcheck disable=SC2086
-        curl -sL https://raw.githubusercontent.com/planetary-annihilation/papatcher/master/papatcher.go -o /usr/local/lanordie/gameserver/papatcher.go
+        curl -sL https://raw.githubusercontent.com/planetary-annihilation/papatcher/master/papatcher.go -o "${INSTALLER_DIR_PATH}"/papatcher.go
         chmod +x papatcher.go
-        go run /usr/local/lanordie/gameserver/papatcher.go \
+        go run "${INSTALLER_DIR_PATH}"/papatcher.go \
             --dir /home/ubuntu/pa_titans/pa/ \
             --stream stable \
             --update-only \
@@ -440,6 +467,8 @@ then
     echo "INFO: Satisfactory Experimental service status"
     systemctl status satisfactory_experimental.service
 fi
+
+# umount the S3 drive
 
 # create cron job to update and restart services once a week