+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.0.22
 
 - MOVED title install / system service creation out of user-data.sh, now resides in installer.sh which is stored in the S3 bucket `installers`.
-- ADDED archive for Factorio, KSP, and PA:Titans to `installer` S3 bucket.
+- ADDED archive for Factorio, KSP, and PA:Titans to `installer` S3 bucket. Titles now install from archives in S3 if missing on EC2 instance
 
 ## 0.0.20